Majlesyar/backend/requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.16,<3.17
django-cors-headers>=4.9,<5.0
djangorestframework-simplejwt>=5.5,<5.6
drf-spectacular>=0.29,<0.30
django-unfold>=0.79,<0.80
django-filter>=25.2,<26.0
whitenoise>=6.9,<7.0
Pillow>=12.1,<13.0
pillow-avif-plugin>=1.5.2,<2.0
psycopg[binary]>=3.3,<4.0
gunicorn>=25.0,<26.0
open_clip_torch>=2.24,<3.0
torch>=2.2,<3.0
torchvision>=0.17,<1.0
timm>=0.9,<1.0
ftfy>=6.2,<7.0
